upload examples of planned schedule (JSON format)
5.1 Planned construction schedule in JSON format
To be compatible with our system, JSON format is used to express each activity information in the planned construction schedule, such as ActivityType, StartDate, Duration, DelayDays, and its Predecessors. To distinguish same activities on different floors and same activities on the same floor but in different construction zones, we use “Construction type name + Floor number + Material Delivery/task name + Zone number” to name the activity to ensure that it is unique and not duplicated with other activities, thus avoiding confusion and errors.
This figure presents one part of the work conducted by the authors.